# openapi.yml
openapi: 3.0.3
info:
  title: buff
  description: ""
  license:
    name: ""
  version: 0.1.0
paths:
...
  /api/admin/moderator/orders:
    get:
      tags:
        - handlers::admin::moderators
      operationId: list_moderators_orders
      responses:
        "200":
          description: Orders were successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "404":
          description: Moderator was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_admin: []
...
  /api/admin/order/{id}/cancel:
    patch:
      tags:
        - handlers::admin::orders
      operationId: cancel_order_by_id
      parameters:
        - name: id
          in: path
          description: Order id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Order was successfully canceled
        "400":
          description: Order has already been marked as succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "404":
          description: Order was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_admin: []
...
  /api/user/order:
    get:
      tags:
        - handlers::orders
      operationId: list_orders
      responses:
        "200":
          description: Orders were successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_user: []
    post:
      tags:
        - handlers::orders
      operationId: create_order
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
        required: true
      responses:
        "201":
          description: Order was successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "404":
          description: Currency symbol was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_user: []
  /api/user/order/{id}:
    get:
      tags:
        - handlers::orders
      operationId: get_order
      parameters:
        - name: id
          in: path
          description: Order id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Order was successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "404":
          description: Order was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_user: []
    patch:
      tags:
        - handlers::orders
      operationId: set_requisites
      parameters:
        - name: id
          in: path
          description: Order id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetRequisitesRequest"
        required: true
      responses:
        "204":
          description: Requisites were successfully set
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "404":
          description: Order was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_user: []
  /api/user/order/{id}/cancel:
    patch:
      tags:
        - handlers::orders
      operationId: cancel_order
      parameters:
        - name: id
          in: path
          description: Order id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Order was successfully canceled
        "400":
          description: Order has already been marked as succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "404":
          description: Order was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_user: []
...
components:
  schemas:
...
    CreateOrderRequest:
      type: object
      required:
        - payment_method
        - amount
        - currency
      properties:
        payment_method:
          type: string
        amount:
          type: string
        currency:
          type: string
...
    Details:
      type: object
      required:
        - details
      properties:
        details:
          type: string
...
    Order:
      type: object
      required:
        - id
        - payment_method
        - status
        - created_at
        - steam_id
        - amount
        - fixed_currency_rate
        - currency_symbol
      properties:
        id:
          type: integer
          format: int64
        payment_method:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        steam_id:
          type: integer
          format: int64
        moderator_id:
          type: integer
          format: int64
          nullable: true
        amount:
          type: string
        fixed_currency_rate:
          type: string
        currency_symbol:
          type: string
        finished_at:
          type: string
          format: date-time
          nullable: true
...
    SetRequisitesRequest:
      type: object
      properties:
        requisites:
          type: string
          nullable: true
...
  securitySchemes:
    jwt_admin:
      type: apiKey
      in: header
      name: X-AM-Authorization
    jwt_user:
      type: http
      scheme: bearer
      bearerFormat: JWT